requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-lru==1.0.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Cython==0.29.28
databases==0.5.5
decorator==5.1.1
fastapi==0.75.0
fire==0.4.0
frequency-controller==0.0.6
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.0.3
jmespath==0.10.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
lxml==4.8.0
MarkupSafe==2.1.0
multidict==6.0.2
objectpath==0.6.1
pbr==5.8.1
ply==3.11
pydantic==1.9.0
PyYAML==6.0
requests==2.27.1
selectolax==0.3.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
starlette==0.17.1
termcolor==1.1.0
toml==0.10.2
torequests==5.1.5
typing_extensions==4.1.1
uniparser==2.0.0
urllib3==1.26.8
uvicorn==0.17.5
watchdog==2.1.6
watchdogs==1.8.8
yarl==1.7.2

coloredlogs~=15.0.1